refactor: Adjust file-names to reflect single entry-point dev pattern
