+++ src/components/Admin/index.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from "react";
 import { useForm } from "react-hook-form";
 import useLogin from "@/components/Admin/UseLogin";
-import SocialMediaIcons from "@/components/SocialMediaIcons/SocialMediaIcons";
+import SocialMediaIcons from "@/components/SocialMediaIcons";
 
 interface LoginFormProps {
   onSubmit: (data: LoginFormInputs) => void;
