+++ src/components/BaseLayout/index.tsx
@@ -1,4 +1,4 @@
-import Navbar from "@/components/Navbar/Navbar";
+import Navbar from "@/components/Navbar";
 interface BaseLayoutProps {
   children: React.ReactNode;
 }
