+++ src/pages/about/index.tsx
@@ -1,4 +1,4 @@
-import About from "@/components/About/About";
+import About from "@/components/About";
 
 export default function AboutPage() {
   return <About />;
