+++ src/pages/admin/index.tsx
@@ -1,4 +1,4 @@
-import Admin from "@/components/Admin/Admin";
+import Admin from "@/components/Admin";
 
 export default function AdminPage() {
   return <Admin />;
