+++ src/pages/help/index.tsx
@@ -1,4 +1,4 @@
-import Help from "@/components/Help/Help";
+import Help from "@/components/Help";
 
 export default function HelpPage() {
   return <Help />;
