+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import Home from "@/components/Home/Home";
+import Home from "@/components/Home";
 
 export default function HomePage() {
   return <Home />;
